チャンネルごとにDocumentを削除するように変更

+++ functions/crawlchannel/src/GCP/Firestore.php
@@ -2,8 +2,10 @@
 namespace F122apg\YoutubeLiveChecker\GCP;
 
 use Google\Cloud\Firestore\FirestoreClient;
+use Google\Cloud\Firestore\QuerySnapshot;
 
 use F122apg\YoutubeLiveChecker\App;
+use F122apg\YoutubeLiveChecker\Log;
 use F122apg\YoutubeLiveChecker\GCP\FeedDocument;
 
 class Firestore {
@@ -26,7 +28,7 @@ class Firestore {
      *
      * @var array
      */
-    private array $cachedDocuments = [];
+    private QuerySnapshot $cachedDocuments;
 
     public function __construct() {
         $this->_client = new FirestoreClient([
@@ -48,15 +50,17 @@ class Firestore {
     }
 
     /**
-     * FirestoreのDocumentを最新（15個）以外削除する
+     * FirestoreのDocumentをチャンネルごとに最新（15個）以外削除する
      *
+     * @param string $channelId チャンネルID
      * @param array $feedContentIds RSSフィードから取得したContentID
      * @return void
      */
-    public function deleteDocuments($feedContentIds) {
+    public function deleteDocuments($channelId, $feedContentIds) {
         foreach ($this->cachedDocuments as $doc) {
             // RSSフィードにないContentIDはもう使用しない古いContentIDなので削除する
-            if (!in_array($doc['contentId'], $feedContentIds, true)) {
+            if ($doc['channelId'] === $channelId && !in_array($doc['contentId'], $feedContentIds, true)) {
+                Log::info('Delete[' . self::_COLLECTION_NAME . '/' . $doc['contentId'] . ']');
                 $this->_client->document(self::_COLLECTION_NAME . '/' . $doc['contentId'])->delete();
             }
         }
@@ -70,10 +74,14 @@ class Firestore {
      */
     public function getNewContentIds(array $contentIds):array {
         $col = $this->_getCollection();
+        $docs = [];
         // キャッシュがあれば、そちらのDocumentsを見る
-        $docs = count($this->cachedDocuments)
-            ? $this->cachedDocuments
-            : $col->documents();
+        if (isset($this->cachedDocuments) && !$this->cachedDocuments->isEmpty()) {
+            $docs = $this->cachedDocuments;
+        } else {
+            $docs = $col->documents();
+            $this->cachedDocuments = $docs;
+        }
 
         if (empty($docs)) {
             return $contentIds;
